Refactor github_manager.py into individual function files
Split services/github/github_manager.py into separate files:
- add_issue_templates → services/github/templates/add_issue_templates.py
- get_latest_remote_commit_sha → services/github/commits/get_latest_remote_commit_sha.py
- get_remote_file_content → services/github/files/get_remote_file_content.py
- get_remote_file_content_by_url → services/github/files/get_remote_file_content_by_url.py
- search_remote_file_contents → services/github/search/search_remote_file_contents.py

Updated all import statements across the codebase to use the new individual function imports.
