--- services/github/github_manager.py
@@ -1,315 +1,0 @@
-# Standard imports
-import base64
-import json
-import logging
-from typing import Optional
-from uuid import uuid4
-
-# Third-party imports
-import requests
-from github import Github
-from github.ContentFile import ContentFile
-from github.PullRequest import PullRequest
-from github.Repository import Repository
-
-# Local imports
-from config import (
-    GITHUB_API_URL,
-    GITHUB_ISSUE_DIR,
-    GITHUB_ISSUE_TEMPLATES,
-    PRODUCT_NAME,
-    TIMEOUT,
-    PRODUCT_ID,
-    UTF8,
-)
-
-# Local imports (GitHub)
-from services.github.comments.update_comment import update_comment
-from services.github.repositories.initialize_repo import initialize_repo
-from services.github.utils.create_headers import create_headers
-from services.github.types.github_types import BaseArgs
-
-# Local imports (OpenAI & Supabase)
-from services.openai.vision import describe_image
-
-# Local imports (Utils)
-from utils.error.handle_exceptions import handle_exceptions
-from utils.files.get_file_content import get_file_content
-from utils.new_lines.detect_new_line import detect_line_break
-from utils.urls.parse_urls import parse_github_url
-
-
-@handle_exceptions(default_return_value=None, raise_on_error=False)
-def add_issue_templates(full_name: str, installer_name: str, token: str) -> None:
-    print(f"Adding issue templates to the repo: '{full_name}' by '{installer_name}'.\n")
-    gh = Github(login_or_token=token)
-    repo: Repository = gh.get_repo(full_name_or_id=full_name)
-    owner, repo_name = full_name.split("/")
-    default_branch_name = repo.default_branch
-
-    # Get latest commit SHA directly
-    base_args = {
-        "owner": owner,
-        "repo": repo_name,
-        "base_branch": default_branch_name,
-        "token": token,
-    }
-    clone_url = repo.clone_url
-    latest_sha = get_latest_remote_commit_sha(clone_url=clone_url, base_args=base_args)
-
-    # Create a new branch using the SHA
-    new_branch_name: str = f"{PRODUCT_ID}/add-issue-templates-{str(object=uuid4())}"
-    ref = f"refs/heads/{new_branch_name}"
-    repo.create_git_ref(ref=ref, sha=latest_sha)
-
-    # Add issue templates to the new branch
-    added_templates: list[str] = []
-    for template_file in GITHUB_ISSUE_TEMPLATES:
-        # Get an issue template content from GitAuto repository
-        template_path: str = GITHUB_ISSUE_DIR + "/" + template_file
-        content = get_file_content(file_path=template_path)
-
-        # Get the list of existing files in the user's remote repository at the GITHUB_ISSUE_DIR. We need to use try except as repo.get_contents() raises a 404 error if the directory doesn't exist. Also directory path MUST end without a slash.
-        try:
-            remote_files: list[ContentFile] = repo.get_contents(path=GITHUB_ISSUE_DIR)
-            remote_file_names: list[str] = [file.name for file in remote_files]
-        except Exception:  # pylint: disable=broad-except
-            remote_file_names = []
-
-        # Skip if the template already exists
-        if template_file in remote_file_names:
-            continue
-
-        # Add file to the new branch
-        msg = f"Add a template: {template_file}"
-        repo.create_file(
-            path=template_path, message=msg, content=content, branch=new_branch_name
-        )
-        added_templates.append(template_file)
-
-    # Return early if no templates were added
-    if not added_templates:
-        return
-
-    # If there are added templates, create a PR
-    pr: PullRequest = repo.create_pull(
-        base=default_branch_name,
-        head=new_branch_name,
-        # Add X issue templates: bug_report.yml, feature_request.yml
-        title=f"Add {len(added_templates)} issue templates",
-        body=f"## Overview\n\nThis PR adds issue templates to the repository so that you can create issues more easily for {PRODUCT_NAME} and your project. Please review the changes and merge the PR if you agree.\n\n## Added templates:\n\n- "
-        + "\n- ".join(added_templates),
-        maintainer_can_modify=True,
-        draft=False,
-    )
-
-    # Add reviewers to the PR. When I tried to add reviewers, I got a 422 error. So, "reviewers" parameter must be an array of strings, not a string.
-    pr.create_review_request(reviewers=[installer_name])
-
-
-@handle_exceptions(raise_on_error=True)
-def get_latest_remote_commit_sha(clone_url: str, base_args: BaseArgs) -> str:
-    """SHA stands for Secure Hash Algorithm. It's a unique identifier for a commit.
-    https://docs.github.com/en/rest/git/refs?apiVersion=2022-11-28#get-a-reference"""
-    owner, repo, branch = (
-        base_args["owner"],
-        base_args["repo"],
-        base_args["base_branch"],
-    )
-    token = base_args["token"]
-    try:
-        response: requests.Response = requests.get(
-            url=f"{GITHUB_API_URL}/repos/{owner}/{repo}/git/ref/heads/{branch}",
-            headers=create_headers(token=token),
-            timeout=TIMEOUT,
-        )
-        response.raise_for_status()
-        return response.json()["object"]["sha"]
-    except requests.exceptions.HTTPError as e:
-        if (
-            e.response.status_code == 409
-            and e.response.json()["message"] == "Git Repository is empty."
-        ):
-            msg = "Repository is empty. So, creating an initial empty commit."
-            logging.info(msg)
-            repo_path = f"/tmp/repo/{owner}-{repo}"
-            initialize_repo(repo_path=repo_path, remote_url=clone_url, token=token)
-            return get_latest_remote_commit_sha(
-                clone_url=clone_url, base_args=base_args
-            )
-        raise
-    except Exception as e:
-        msg = f"{get_latest_remote_commit_sha.__name__} encountered an error: {e}"
-        update_comment(body=msg, base_args=base_args)
-        # Raise an error because we can't continue without the latest commit SHA
-        raise RuntimeError(
-            f"Error: Could not get the latest commit SHA in {get_latest_remote_commit_sha.__name__}"
-        ) from e
-
-
-@handle_exceptions(default_return_value="", raise_on_error=False)
-def get_remote_file_content(
-    file_path: str,
-    base_args: BaseArgs,
-    line_number: Optional[int] = None,
-    keyword: Optional[str] = None,
-    **_kwargs,
-) -> str:
-    """
-    https://docs.github.com/en/rest/repos/contents?apiVersion=2022-11-28
-
-    params:
-    - file_path: file path or directory path. Ex) 'src/main.py' or 'src'
-    - line_number: specific line number to focus on
-    - keyword: keyword to search in the file content
-    """
-    if line_number is not None and keyword is not None:
-        return "Error: You can only specify either line_number or keyword, not both."
-
-    # Convert line_number to int if it's a string
-    if line_number is not None and isinstance(line_number, str):
-        try:
-            line_number = int(line_number)
-        except ValueError:
-            return f"Error: line_number '{line_number}' is not a valid integer."
-
-    owner, repo, ref, token = (
-        base_args["owner"],
-        base_args["repo"],
-        base_args["new_branch"],
-        base_args["token"],
-    )
-    url = f"{GITHUB_API_URL}/repos/{owner}/{repo}/contents/{file_path}?ref={ref}"
-    headers: dict[str, str] = create_headers(token=token)
-    response = requests.get(url=url, headers=headers, timeout=TIMEOUT)
-
-    # If 404 error, return early. Otherwise, raise a HTTPError
-    if response.status_code == 404:
-        return f"{get_remote_file_content.__name__} encountered an HTTPError: 404 Client Error: Not Found for url: {url}. Check the file path, correct it, and try again."
-    response.raise_for_status()
-
-    # file_path is expected to be a file path, but it can be a directory path due to AI's volatility. See Example2 at https://docs.github.com/en/rest/repos/contents?apiVersion=2022-11-28
-    res_json = response.json()
-    if not isinstance(res_json, dict):
-        file_paths: list[str] = [item["path"] for item in res_json]
-        msg = f"Searched directory '{file_path}' and found: {json.dumps(file_paths)}"
-        return msg
-
-    encoded_content: str = res_json["content"]  # Base64 encoded content
-
-    # If encoded_content is image, describe the image content in text by vision API
-    if file_path.endswith((".png", ".jpeg", ".jpg", ".webp", ".gif")):
-        msg = f"Opened image file: '{file_path}' and described the content.\n\n"
-        return msg + describe_image(base64_image=encoded_content)
-
-    # Otherwise, decode the content
-    decoded_content: str = base64.b64decode(s=encoded_content).decode(encoding=UTF8)
-    lb: str = detect_line_break(text=decoded_content)
-    lines = decoded_content.split(lb)
-    width = len(str(len(lines)))
-    numbered_lines = [f"{i + 1:>{width}}:{line}" for i, line in enumerate(lines)]
-    file_path_with_lines = file_path
-
-    # If line_number is specified, show the lines around the line_number
-    buffer = 50
-    if line_number is not None and line_number > 1 and len(lines) > 100:
-        start = max(line_number - buffer, 0)
-        end = min(line_number + buffer, len(lines))
-        numbered_lines = numbered_lines[start : end + 1]  # noqa: E203
-        file_path_with_lines = f"{file_path}#L{start + 1}-L{end + 1}"
-
-    # If keyword is specified, show the lines containing the keyword
-    elif keyword is not None:
-        segments = []
-        for i, line in enumerate(lines):
-            if keyword not in line:
-                continue
-            start = max(i - buffer, 0)
-            end = min(i + buffer, len(lines))
-            segment = lb.join(numbered_lines[start : end + 1])  # noqa: E203
-            file_path_with_lines = f"{file_path}#L{start + 1}-L{end + 1}"
-            segments.append(f"```{file_path_with_lines}\n" + segment + "\n```")
-
-        if not segments:
-            return f"Keyword '{keyword}' not found in the file '{file_path}'."
-        msg = f"Opened file: '{file_path}' and found multiple occurrences of '{keyword}'.\n\n"
-        return msg + "\n\n•\n•\n•\n\n".join(segments)
-
-    numbered_content: str = lb.join(numbered_lines)
-    msg = f"Opened file: '{file_path}' with line numbers for your information.\n\n"
-    return msg + f"```{file_path_with_lines}\n{numbered_content}\n```"
-
-
-@handle_exceptions(default_return_value="", raise_on_error=False)
-def get_remote_file_content_by_url(url: str, token: str) -> str:
-    """https://docs.github.com/en/rest/repos/contents?apiVersion=2022-11-28"""
-    parts = parse_github_url(url)
-    owner, repo, ref, file_path = (
-        parts["owner"],
-        parts["repo"],
-        parts["ref"],
-        parts["file_path"],
-    )
-    start, end = parts["start_line"], parts["end_line"]
-    url: str = f"{GITHUB_API_URL}/repos/{owner}/{repo}/contents/{file_path}?ref={ref}"
-    headers: dict[str, str] = create_headers(token=token)
-    response = requests.get(url=url, headers=headers, timeout=TIMEOUT)
-    response.raise_for_status()
-    response_json = response.json()
-    encoded_content: str = response_json["content"]  # Base64 encoded content
-    decoded_content: str = base64.b64decode(s=encoded_content).decode(encoding=UTF8)
-    numbered_lines = [
-        f"{i + 1}: {line}" for i, line in enumerate(decoded_content.split("\n"))
-    ]
-
-    if start is not None and end is not None:
-        numbered_lines = numbered_lines[start - 1 : end]  # noqa: E203
-        file_path_with_lines = f"{file_path}#L{start}-L{end}"
-    elif start is not None:
-        numbered_lines = numbered_lines[start - 1]  # noqa: E203
-        file_path_with_lines = f"{file_path}#L{start}"
-    else:
-        file_path_with_lines = file_path
-
-    numbered_content: str = "\n".join(numbered_lines)
-    return f"## {file_path_with_lines}\n\n{numbered_content}"
-
-
-@handle_exceptions(default_return_value="", raise_on_error=False)
-def search_remote_file_contents(query: str, base_args: BaseArgs, **_kwargs) -> str:
-    """
-    - Only the default branch is considered.
-    - Only files smaller than 384 KB are searchable.
-    - This endpoint requires you to authenticate and limits you to 10 requests per minute.
-
-    https://docs.github.com/en/rest/search/search?apiVersion=2022-11-28
-    https://docs.github.com/en/search-github/getting-started-with-searching-on-github/understanding-the-search-syntax
-    https://docs.github.com/en/search-github/searching-on-github/searching-in-forks
-    """
-    owner, repo, is_fork, token = (
-        base_args["owner"],
-        base_args["repo"],
-        base_args["is_fork"],
-        base_args["token"],
-    )
-    q = f"{query} repo:{owner}/{repo}"
-    if is_fork:
-        q = f"{query} repo:{owner}/{repo} fork:true"
-    params = {"q": q, "per_page": 10, "page": 1}  # per_page: max 100
-    url = f"{GITHUB_API_URL}/search/code"
-    headers: dict[str, str] = create_headers(token=token)
-    headers["Accept"] = "application/vnd.github.text-match+json"
-    response = requests.get(url=url, headers=headers, params=params, timeout=TIMEOUT)
-    response.raise_for_status()
-    response_json = response.json()
-    file_paths = []
-    for item in response_json.get("items", []):
-        file_path = item["path"]
-        file_paths.append(file_path)
-    msg = (
-        f"{len(file_paths)} files found for the search query '{query}':\n- "
-        + "\n- ".join(file_paths)
-        + "\n"
-    )
-    print(msg)
-    return msg
