+++ services/github/issues/test_is_issue_open.py
@@ -1,5 +1,4 @@
 from unittest.mock import patch, MagicMock
-import requests
 from requests.exceptions import HTTPError, JSONDecodeError, Timeout, ConnectionError
 
 from services.github.issues.is_issue_open import is_issue_open
@@ -9,18 +8,19 @@ from tests.constants import TOKEN
 def test_is_issue_open_with_open_issue():
     """Test is_issue_open returns True for an open issue."""
     issue_url = "https://github.com/owner/repo/issues/123"
-    
-    with patch("services.github.issues.is_issue_open.requests.get") as mock_get, \
-         patch("services.github.issues.is_issue_open.create_headers") as mock_create_headers:
-        
+
+    with patch("services.github.issues.is_issue_open.requests.get") as mock_get, patch(
+        "services.github.issues.is_issue_open.create_headers"
+    ) as mock_create_headers:
+
         mock_response = MagicMock()
         mock_response.status_code = 200
         mock_response.json.return_value = {"state": "open"}
         mock_get.return_value = mock_response
         mock_create_headers.return_value = {"Authorization": f"Bearer {TOKEN}"}
-        
+
         result = is_issue_open(issue_url, TOKEN)
-        
+
         assert result is True
         mock_get.assert_called_once()
         mock_response.json.assert_called_once()
@@ -29,18 +29,19 @@ def test_is_issue_open_with_open_issue():
 def test_is_issue_open_with_closed_issue():
     """Test is_issue_open returns False for a closed issue."""
     issue_url = "https://github.com/owner/repo/issues/456"
-    
-    with patch("services.github.issues.is_issue_open.requests.get") as mock_get, \
-         patch("services.github.issues.is_issue_open.create_headers") as mock_create_headers:
-        
+
+    with patch("services.github.issues.is_issue_open.requests.get") as mock_get, patch(
+        "services.github.issues.is_issue_open.create_headers"
+    ) as mock_create_headers:
+
         mock_response = MagicMock()
         mock_response.status_code = 200
         mock_response.json.return_value = {"state": "closed"}
         mock_get.return_value = mock_response
         mock_create_headers.return_value = {"Authorization": f"Bearer {TOKEN}"}
-        
+
         result = is_issue_open(issue_url, TOKEN)
-        
+
         assert result is False
         mock_get.assert_called_once()
         mock_response.json.assert_called_once()
@@ -50,7 +51,7 @@ def test_is_issue_open_with_empty_url():
     """Test is_issue_open returns False for empty URL."""
     result = is_issue_open("", TOKEN)
     assert result is False
-    
+
     result = is_issue_open(None, TOKEN)
     assert result is False
 
@@ -64,7 +65,7 @@ def test_is_issue_open_with_invalid_url_format():
         "https://example.com/owner/repo/issues/123",  # Not GitHub
         "not-a-url-at-all",  # Invalid URL
     ]
-    
+
     for url in invalid_urls:
         result = is_issue_open(url, TOKEN)
         assert result is True, f"Expected True for invalid URL: {url}"
@@ -73,20 +74,24 @@ def test_is_issue_open_with_invalid_url_format():
 def test_is_issue_open_api_url_construction():
     """Test that the correct GitHub API URL is constructed."""
     issue_url = "https://github.com/test-owner/test-repo/issues/789"
-    
-    with patch("services.github.issues.is_issue_open.requests.get") as mock_get, \
-         patch("services.github.issues.is_issue_open.create_headers") as mock_create_headers, \
-         patch("services.github.issues.is_issue_open.GITHUB_API_URL", "https://api.github.com"):
-        
+
+    with patch("services.github.issues.is_issue_open.requests.get") as mock_get, patch(
+        "services.github.issues.is_issue_open.create_headers"
+    ) as mock_create_headers, patch(
+        "services.github.issues.is_issue_open.GITHUB_API_URL", "https://api.github.com"
+    ):
+
         mock_response = MagicMock()
         mock_response.status_code = 200
         mock_response.json.return_value = {"state": "open"}
         mock_get.return_value = mock_response
         mock_create_headers.return_value = {"Authorization": f"Bearer {TOKEN}"}
-        
+
         is_issue_open(issue_url, TOKEN)
-        
-        expected_api_url = "https://api.github.com/repos/test-owner/test-repo/issues/789"
+
+        expected_api_url = (
+            "https://api.github.com/repos/test-owner/test-repo/issues/789"
+        )
         mock_get.assert_called_once()
         call_args = mock_get.call_args
         assert call_args.kwargs["url"] == expected_api_url
@@ -96,19 +101,20 @@ def test_is_issue_open_headers_creation():
     """Test that correct headers are created and passed."""
     issue_url = "https://github.com/owner/repo/issues/123"
     test_token = "test-token-123"
-    
-    with patch("services.github.issues.is_issue_open.requests.get") as mock_get, \
-         patch("services.github.issues.is_issue_open.create_headers") as mock_create_headers:
-        
+
+    with patch("services.github.issues.is_issue_open.requests.get") as mock_get, patch(
+        "services.github.issues.is_issue_open.create_headers"
+    ) as mock_create_headers:
+
         mock_response = MagicMock()
         mock_response.status_code = 200
         mock_response.json.return_value = {"state": "open"}
         mock_get.return_value = mock_response
         expected_headers = {"Authorization": f"Bearer {test_token}"}
         mock_create_headers.return_value = expected_headers
-        
+
         is_issue_open(issue_url, test_token)
-        
+
         mock_create_headers.assert_called_once_with(token=test_token)
         call_args = mock_get.call_args
         assert call_args.kwargs["headers"] == expected_headers
@@ -117,19 +123,19 @@ def test_is_issue_open_headers_creation():
 def test_is_issue_open_timeout_parameter():
     """Test that timeout parameter is correctly passed."""
     issue_url = "https://github.com/owner/repo/issues/123"
-    
-    with patch("services.github.issues.is_issue_open.requests.get") as mock_get, \
-         patch("services.github.issues.is_issue_open.create_headers") as mock_create_headers, \
-         patch("services.github.issues.is_issue_open.TIMEOUT", 60):
-        
+
+    with patch("services.github.issues.is_issue_open.requests.get") as mock_get, patch(
+        "services.github.issues.is_issue_open.create_headers"
+    ) as mock_create_headers, patch("services.github.issues.is_issue_open.TIMEOUT", 60):
+
         mock_response = MagicMock()
         mock_response.status_code = 200
         mock_response.json.return_value = {"state": "open"}
         mock_get.return_value = mock_response
         mock_create_headers.return_value = {"Authorization": f"Bearer {TOKEN}"}
-        
+
         is_issue_open(issue_url, TOKEN)
-        
+
         call_args = mock_get.call_args
         assert call_args.kwargs["timeout"] == 60
 
@@ -137,20 +143,23 @@ def test_is_issue_open_timeout_parameter():
 def test_is_issue_open_non_200_status_code():
     """Test is_issue_open returns True for non-200 status codes (assumes open to avoid duplicates)."""
     issue_url = "https://github.com/owner/repo/issues/123"
-    
+
     status_codes = [404, 403, 500, 502, 503]
-    
+
     for status_code in status_codes:
-        with patch("services.github.issues.is_issue_open.requests.get") as mock_get, \
-             patch("services.github.issues.is_issue_open.create_headers") as mock_create_headers:
-            
+        with patch(
+            "services.github.issues.is_issue_open.requests.get"
+        ) as mock_get, patch(
+            "services.github.issues.is_issue_open.create_headers"
+        ) as mock_create_headers:
+
             mock_response = MagicMock()
             mock_response.status_code = status_code
             mock_get.return_value = mock_response
             mock_create_headers.return_value = {"Authorization": f"Bearer {TOKEN}"}
-            
+
             result = is_issue_open(issue_url, TOKEN)
-            
+
             assert result is True, f"Expected True for status code {status_code}"
             mock_get.assert_called_once()
             # json() should not be called for non-200 status codes
@@ -160,18 +169,19 @@ def test_is_issue_open_non_200_status_code():
 def test_is_issue_open_missing_state_in_response():
     """Test is_issue_open returns False when state is missing from response."""
     issue_url = "https://github.com/owner/repo/issues/123"
-    
-    with patch("services.github.issues.is_issue_open.requests.get") as mock_get, \
-         patch("services.github.issues.is_issue_open.create_headers") as mock_create_headers:
-        
+
+    with patch("services.github.issues.is_issue_open.requests.get") as mock_get, patch(
+        "services.github.issues.is_issue_open.create_headers"
+    ) as mock_create_headers:
+
         mock_response = MagicMock()
         mock_response.status_code = 200
         mock_response.json.return_value = {}  # Missing 'state' key
         mock_get.return_value = mock_response
         mock_create_headers.return_value = {"Authorization": f"Bearer {TOKEN}"}
-        
+
         result = is_issue_open(issue_url, TOKEN)
-        
+
         assert result is False
         mock_get.assert_called_once()
         mock_response.json.assert_called_once()
@@ -180,21 +190,24 @@ def test_is_issue_open_missing_state_in_response():
 def test_is_issue_open_state_not_open():
     """Test is_issue_open returns False for states other than 'open'."""
     issue_url = "https://github.com/owner/repo/issues/123"
-    
+
     non_open_states = ["closed", "draft", "merged", "unknown", None, ""]
-    
+
     for state in non_open_states:
-        with patch("services.github.issues.is_issue_open.requests.get") as mock_get, \
-             patch("services.github.issues.is_issue_open.create_headers") as mock_create_headers:
-            
+        with patch(
+            "services.github.issues.is_issue_open.requests.get"
+        ) as mock_get, patch(
+            "services.github.issues.is_issue_open.create_headers"
+        ) as mock_create_headers:
+
             mock_response = MagicMock()
             mock_response.status_code = 200
             mock_response.json.return_value = {"state": state}
             mock_get.return_value = mock_response
             mock_create_headers.return_value = {"Authorization": f"Bearer {TOKEN}"}
-            
+
             result = is_issue_open(issue_url, TOKEN)
-            
+
             assert result is False, f"Expected False for state: {state}"
             mock_get.assert_called_once()
             mock_response.json.assert_called_once()
@@ -203,10 +216,11 @@ def test_is_issue_open_state_not_open():
 def test_is_issue_open_http_error():
     """Test is_issue_open returns True when HTTPError occurs (assumes open to avoid duplicates)."""
     issue_url = "https://github.com/owner/repo/issues/123"
-    
-    with patch("services.github.issues.is_issue_open.requests.get") as mock_get, \
-         patch("services.github.issues.is_issue_open.create_headers") as mock_create_headers:
-        
+
+    with patch("services.github.issues.is_issue_open.requests.get") as mock_get, patch(
+        "services.github.issues.is_issue_open.create_headers"
+    ) as mock_create_headers:
+
         mock_response = MagicMock()
         mock_response.status_code = 404
         mock_response.reason = "Not Found"
@@ -216,14 +230,14 @@ def test_is_issue_open_http_error():
             "X-RateLimit-Remaining": "4999",
             "X-RateLimit-Used": "1",
         }
-        
+
         http_error = HTTPError("404 Client Error")
         http_error.response = mock_response
         mock_get.side_effect = http_error
         mock_create_headers.return_value = {"Authorization": f"Bearer {TOKEN}"}
-        
+
         result = is_issue_open(issue_url, TOKEN)
-        
+
         assert result is True
         mock_get.assert_called_once()
 
@@ -231,18 +245,19 @@ def test_is_issue_open_http_error():
 def test_is_issue_open_json_decode_error():
     """Test is_issue_open returns True when JSONDecodeError occurs."""
     issue_url = "https://github.com/owner/repo/issues/123"
-    
-    with patch("services.github.issues.is_issue_open.requests.get") as mock_get, \
-         patch("services.github.issues.is_issue_open.create_headers") as mock_create_headers:
-        
+
+    with patch("services.github.issues.is_issue_open.requests.get") as mock_get, patch(
+        "services.github.issues.is_issue_open.create_headers"
+    ) as mock_create_headers:
+
         mock_response = MagicMock()
         mock_response.status_code = 200
         mock_response.json.side_effect = JSONDecodeError("Invalid JSON", "{", 0)
         mock_get.return_value = mock_response
         mock_create_headers.return_value = {"Authorization": f"Bearer {TOKEN}"}
-        
+
         result = is_issue_open(issue_url, TOKEN)
-        
+
         assert result is True
         mock_get.assert_called_once()
         mock_response.json.assert_called_once()
@@ -251,15 +266,16 @@ def test_is_issue_open_json_decode_error():
 def test_is_issue_open_timeout_error():
     """Test is_issue_open returns True when Timeout occurs."""
     issue_url = "https://github.com/owner/repo/issues/123"
-    
-    with patch("services.github.issues.is_issue_open.requests.get") as mock_get, \
-         patch("services.github.issues.is_issue_open.create_headers") as mock_create_headers:
-        
+
+    with patch("services.github.issues.is_issue_open.requests.get") as mock_get, patch(
+        "services.github.issues.is_issue_open.create_headers"
+    ) as mock_create_headers:
+
         mock_get.side_effect = Timeout("Request timed out")
         mock_create_headers.return_value = {"Authorization": f"Bearer {TOKEN}"}
-        
+
         result = is_issue_open(issue_url, TOKEN)
-        
+
         assert result is True
         mock_get.assert_called_once()
 
@@ -267,15 +283,16 @@ def test_is_issue_open_timeout_error():
 def test_is_issue_open_connection_error():
     """Test is_issue_open returns True when ConnectionError occurs."""
     issue_url = "https://github.com/owner/repo/issues/123"
-    
-    with patch("services.github.issues.is_issue_open.requests.get") as mock_get, \
-         patch("services.github.issues.is_issue_open.create_headers") as mock_create_headers:
-        
+
+    with patch("services.github.issues.is_issue_open.requests.get") as mock_get, patch(
+        "services.github.issues.is_issue_open.create_headers"
+    ) as mock_create_headers:
+
         mock_get.side_effect = ConnectionError("Connection failed")
         mock_create_headers.return_value = {"Authorization": f"Bearer {TOKEN}"}
-        
+
         result = is_issue_open(issue_url, TOKEN)
-        
+
         assert result is True
         mock_get.assert_called_once()
 
@@ -283,18 +300,19 @@ def test_is_issue_open_connection_error():
 def test_is_issue_open_key_error():
     """Test is_issue_open returns True when KeyError occurs."""
     issue_url = "https://github.com/owner/repo/issues/123"
-    
-    with patch("services.github.issues.is_issue_open.requests.get") as mock_get, \
-         patch("services.github.issues.is_issue_open.create_headers") as mock_create_headers:
-        
+
+    with patch("services.github.issues.is_issue_open.requests.get") as mock_get, patch(
+        "services.github.issues.is_issue_open.create_headers"
+    ) as mock_create_headers:
+
         mock_response = MagicMock()
         mock_response.status_code = 200
         mock_response.json.side_effect = KeyError("Missing key")
         mock_get.return_value = mock_response
         mock_create_headers.return_value = {"Authorization": f"Bearer {TOKEN}"}
-        
+
         result = is_issue_open(issue_url, TOKEN)
-        
+
         assert result is True
         mock_get.assert_called_once()
         mock_response.json.assert_called_once()
@@ -311,35 +329,43 @@ def test_is_issue_open_url_parsing_edge_cases():
         "https://github.com/a/b/issues/1",  # Minimal valid case
         "https://github.com/owner/repo/issues/",  # Missing issue number but still makes API call
     ]
-    
+
     # URLs that will NOT make API calls (return True early)
     no_api_call_urls = [
         "https://github.com/owner/repo/issues",  # Missing slash and number - invalid format
     ]
-    
+
     # Test URLs that make API calls
     for url in api_call_urls:
-        with patch("services.github.issues.is_issue_open.requests.get") as mock_get, \
-             patch("services.github.issues.is_issue_open.create_headers") as mock_create_headers:
-            
+        with patch(
+            "services.github.issues.is_issue_open.requests.get"
+        ) as mock_get, patch(
+            "services.github.issues.is_issue_open.create_headers"
+        ) as mock_create_headers:
+
             # Mock successful response
             mock_response = MagicMock()
             mock_response.status_code = 200
             mock_response.json.return_value = {"state": "open"}
             mock_get.return_value = mock_response
             mock_create_headers.return_value = {"Authorization": f"Bearer {TOKEN}"}
-            
+
             result = is_issue_open(url, TOKEN)
             assert result is True, f"Expected True for URL that makes API call: {url}"
             mock_get.assert_called_once()
-    
+
     # Test URLs that do NOT make API calls
     for url in no_api_call_urls:
-        with patch("services.github.issues.is_issue_open.requests.get") as mock_get, \
-             patch("services.github.issues.is_issue_open.create_headers") as mock_create_headers:
-            
+        with patch(
+            "services.github.issues.is_issue_open.requests.get"
+        ) as mock_get, patch(
+            "services.github.issues.is_issue_open.create_headers"
+        ) as mock_create_headers:
+
             result = is_issue_open(url, TOKEN)
-            assert result is True, f"Expected True for URL that doesn't make API call: {url}"
+            assert (
+                result is True
+            ), f"Expected True for URL that doesn't make API call: {url}"
             mock_get.assert_not_called()
 
 
@@ -351,11 +377,11 @@ def test_is_issue_open_decorator_applied():
 def test_is_issue_open_function_signature():
     """Test that the function has the correct signature."""
     import inspect
-    
+
     sig = inspect.signature(is_issue_open)
     params = list(sig.parameters.keys())
-    
+
     assert params == ["issue_url", "token"]
-    assert sig.parameters["issue_url"].annotation == str
-    assert sig.parameters["token"].annotation == str
-    assert sig.return_annotation == bool
+    assert sig.parameters["issue_url"].annotation is str
+    assert sig.parameters["token"].annotation is str
+    assert sig.return_annotation is bool
